Fix bug introduced with source transport upgrade
There were still subtle differences between xgpu and fgpu in the way
that the sources were specified, which I missed with unit testing
because of the mockers.

Imagine my surprise when I tried to build a correlator for qualification
testing and it just crashes and burns. I guess this is why we have
tests.

I've moved parse_source() from fgpu's main to the util module, it seemed
more robust than manually casting the Endpoint to a tuple of str and int.

Closes NGC-595.
